fix(05-01): cast age_months_max to integer in vaccine seed
The age_months_max column is null in every VALUES row, so Postgres typed
the derived column as text and the insert into the integer column failed
(42804). Cast to integer so the physician-approved SUS seed applies cleanly.
Migrations applied to live DB acstugafrgrqzvtuznxv in order: tables, RLS, seed.

+++ supabase/migrations/20260720000200_seed_vaccine_schedules.sql
@@ -30,7 +30,7 @@ where not exists (
 --    Inserido GLOBALMENTE (join no schedule por source/version), UMA VEZ, sem coluna de dono.
 insert into public.vaccine_schedule_items
   (schedule_id, vaccine, dose, age_months, age_months_max, week_min, week_max, age_label, sort_order, notes)
-select s.id, v.vaccine, v.dose, v.age_months, v.age_months_max, null, null, v.age_label, v.sort_order, v.notes
+select s.id, v.vaccine, v.dose, v.age_months, v.age_months_max::integer, null, null, v.age_label, v.sort_order, v.notes
 from (values
   -- Ao nascer
   ('BCG',                          'Dose única',   0,  null, 'Ao nascer',  0,  null),
